feat: graduate XDSCommandPalette to core
git mv all CommandPalette source files from packages/lab to packages/core
so rename history is preserved (git log --follow works).

- packages/core/src/CommandPalette/ — all components moved here
- packages/core/package.json — ./CommandPalette subpath export
- packages/core/src/index.ts — re-exports CommandPalette
- packages/lab/src/CommandPalette/index.ts — @deprecated re-export stub
- packages/lab/src/index.ts — delegates to @xds/core/CommandPalette
- Removed lab-only files (doc.mjs, README.md)
- Storybook: import from @xds/core/CommandPalette, title Lab/ → CommandPalette/

Components promoted:
  XDSCommandPalette, XDSCommandPaletteInput, XDSCommandPaletteItem,
  XDSCommandPaletteGroup, XDSCommandPaletteList, XDSCommandPaletteFooter,
  XDSCommandPaletteEmpty, useCommandPaletteContext
