+++ packages/core/src/CommandPalette/CommandPaletteContext.ts
