+++ packages/core/src/CommandPalette/XDSCommandPaletteEmpty.tsx
