+++ packages/core/src/CommandPalette/XDSCommandPaletteFooter.test.tsx
