+++ packages/core/src/CommandPalette/XDSCommandPaletteFooter.tsx
