+++ packages/core/src/CommandPalette/XDSCommandPaletteGroup.tsx
