+++ packages/core/src/CommandPalette/XDSCommandPaletteItem.tsx
