+++ packages/lab/src/CommandPalette/index.ts
@@ -1,32 +1,24 @@
 /**
- * @file index.ts
- * @input Imports CommandPalette components and types
- * @output Exports all CommandPalette components and types
- * @position Component entry point; re-exported by /packages/lab/src/index.ts
- *
- * SYNC: When modified, update this header
+ * @file CommandPalette barrel export
+ * @deprecated Import from '@xds/core/CommandPalette' instead.
+ * This re-export exists for backwards compatibility.
  */
 
-export {XDSCommandPalette} from './XDSCommandPalette';
-export type {XDSCommandPaletteProps} from './XDSCommandPalette';
-
-export {XDSCommandPaletteInput} from './XDSCommandPaletteInput';
-export type {XDSCommandPaletteInputProps} from './XDSCommandPaletteInput';
-
-export {XDSCommandPaletteList} from './XDSCommandPaletteList';
-export type {XDSCommandPaletteListProps} from './XDSCommandPaletteList';
-
-export {XDSCommandPaletteItem} from './XDSCommandPaletteItem';
-export type {XDSCommandPaletteItemProps} from './XDSCommandPaletteItem';
-
-export {XDSCommandPaletteGroup} from './XDSCommandPaletteGroup';
-export type {XDSCommandPaletteGroupProps} from './XDSCommandPaletteGroup';
-
-export {XDSCommandPaletteFooter} from './XDSCommandPaletteFooter';
-export type {XDSCommandPaletteFooterProps} from './XDSCommandPaletteFooter';
-
-export {XDSCommandPaletteEmpty} from './XDSCommandPaletteEmpty';
-export type {XDSCommandPaletteEmptyProps} from './XDSCommandPaletteEmpty';
-
-export {useCommandPaletteContext} from './CommandPaletteContext';
-export type {CommandPaletteContextValue} from './CommandPaletteContext';
+export {
+  XDSCommandPalette,
+  type XDSCommandPaletteProps,
+  XDSCommandPaletteInput,
+  type XDSCommandPaletteInputProps,
+  XDSCommandPaletteList,
+  type XDSCommandPaletteListProps,
+  XDSCommandPaletteItem,
+  type XDSCommandPaletteItemProps,
+  XDSCommandPaletteGroup,
+  type XDSCommandPaletteGroupProps,
+  XDSCommandPaletteFooter,
+  type XDSCommandPaletteFooterProps,
+  XDSCommandPaletteEmpty,
+  type XDSCommandPaletteEmptyProps,
+  useCommandPaletteContext,
+  type CommandPaletteContextValue,
+} from '@xds/core/CommandPalette';
